+++ changelog.md
@@ -366,5 +366,7 @@
 #2.0.11
 * Refactor
 * Enabled polyfills in test app
-* Added getBlobFromUrl method in XmatRest
-* Added downloadBlobFromUrl
+* Added getBlobFromUrl method in XmatRestService
+* Added downloadBlobFromUrl in XmatRestService
+* New model XmatFile
+* Test App updated